Add an alternative function for buffer reuse
This new function has the same basic name as the one-shot function but
adds a "m" for "multiple".

Other than the original function, a valid buffer pointer and size must
be passed to the function and will be updated.

Also no attempt will be made to ever shrink the buffer.

This function is intended for the use in loops where too frequent
reallocations will be avoided.