Simplify Makefile with inference/.DEFAULT rules
Replace all invocations of the $C macro with a single inference rule.

Replace all invocations of the $L macro with a single .DEFAULT rule.

In order to still specify the target-specific dependencies, create
explicit target rules without commands which collect the dependencies
for the inference rules and the .DEFAULT target.