Replace .DEFAULT rule by .o inference rule
It turned out that the .DEFAULT rules was not used at all! Instead, a
predefined internal inference rule was used, which was sufficient to
link the targets.

However, this rule did not use LDFLAGS in the argument order as I
defined it in the .DEFAULT rule.

Fix this by using a .o inference rule. However, it is unclear how to
reliably obtain the list of object files as linker input. Using $? seems
to work in most situations, but "make" might fail of some of the object
files are selectively deleted manually after building the target
successfully in a previous run.

Nevertheless, calling "make" again seems to build the target correctly,
so this is not a big issue. Also I did not find a better way to get a
list of prerequisites.