Add optimization for copying literal characters
Even though this adds 4 lines to the code, it avoids potentially
expensive memcpy() invocations for the case that a character from the
format string shall be copied literally rather than be interpreted as a
format instruction.

C implementations where memcpy() is implemented as a built-in might not
gain much from this optimization, but implementations where memcpy() is
a normal function call should benefit.